Moved the configuration functions to internal/cfgplugins package
